+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improved solver error messages.
 - `poetry` now always reads/writes the `pyproject.toml` file with the `utf-8` encoding.
 - `config --list` now lists all available settings.
+- `init` no longer adds `pytest` to development dependencies.
 
 ### Fixed